Added Background

+++ gui/main.py
@@ -64,8 +64,8 @@ class LoadingOverlay(QFrame):
         self.resize(self.parent().size())
         # Center spinner_label explicitly
         self.spinner_label.setGeometry(QRect(
-            (self.width() - 64) / 2,
-            (self.height() - 64) / 2,
+            int((self.width() - 64) / 2),
+            int((self.height() - 64) / 2),
             64,
             64
         ))
